alfred/requirements/local.txt
-r testing.txt

Sphinx

django-test-without-migrations
